userbot/requirements.txt
art
rich
setuptools
telethon
PySocks
python-socks
